notify-api/requirements/prod.txt
Flask
Flask-Babel
Flask-Cors
Flask-Migrate
Flask-Pydantic
Flask-SQLAlchemy
flask-jwt-oidc
launchdarkly-server-sdk
pydantic
sentry-sdk[flask]
email-validator
notifications-python-client
beautifulsoup4
phonenumbers